Moved classes to non-default packages, shouldn't break anything. Will merge to master
