--- src/Feat_wizard.java
@@ -1,19 +1,0 @@
-import org.eclipse.swt.widgets.Display;
-import org.eclipse.swt.widgets.Shell;
-
-/**
- * The class that handle Feat wizard interface, input and export.
- * @author Innocentius Shellingford
- *
- */
-public class Feat_wizard {
-	private Shell shell;
-	public Feat_wizard(Display display)
-	{
-		
-	}
-	public Shell getshell()
-	{
-		return shell;
-	}
-}
