+++ src/core/Armor.java
@@ -1,3 +1,4 @@
+package core;
 
 public class Armor extends Item {
