+++ src/core/CharClass.java
@@ -1,3 +1,4 @@
+package core;
 
 public class CharClass {
