+++ src/core/CharSkill.java
@@ -1,3 +1,4 @@
+package core;
 
 public class CharSkill {
 	private Skill skill;
