+++ src/core/DnDie.java
@@ -1,3 +1,4 @@
+package core;
 //NON-BRANCH WORK
 import java.util.Random;
