+++ src/core/Item.java
@@ -1,3 +1,4 @@
+package core;
 
 public class Item {
