+++ src/core/Race.java
@@ -1,3 +1,4 @@
+package core;
 import java.util.ArrayList;
