+++ src/core/Spell.java
@@ -1,3 +1,4 @@
+package core;
 
 public class Spell {
