+++ src/core/Weapon.java
@@ -1,3 +1,4 @@
+package core;
 
 public class Weapon extends Item {
