+++ src/core/xmlLoader.java
@@ -1,3 +1,4 @@
+package core;
 import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Paths;
