+++ src/guis/Feat_wizard.java
@@ -1,0 +1,20 @@
+package guis;
+import org.eclipse.swt.widgets.Display;
+import org.eclipse.swt.widgets.Shell;
+
+/**
+ * The class that handle Feat wizard interface, input and export.
+ * @author Innocentius Shellingford
+ *
+ */
+public class Feat_wizard {
+	private Shell shell;
+	public Feat_wizard(Display display)
+	{
+		
+	}
+	public Shell getshell()
+	{
+		return shell;
+	}
+}
