+++ src/guis/Spell_wizard.java
@@ -1,0 +1,20 @@
+package guis;
+import org.eclipse.swt.widgets.Display;
+import org.eclipse.swt.widgets.Shell;
+
+/**
+ * The class that handle Spell wizard interface, input and export.
+ * @author Innocentius Shellingford
+ *
+ */
+public class Spell_wizard {
+	private Shell shell;
+	public Spell_wizard(Display display)
+	{
+		
+	}
+	public Shell getshell()
+	{
+		return shell;
+	}
+}
