+++ src/guis/TrapInterface.java
@@ -1,3 +1,4 @@
+package guis;
 
 public class TrapInterface {
