+++ src/guis/Wizard.java
@@ -1,3 +1,4 @@
+package guis;
 import org.eclipse.swt.graphics.*;
 import org.eclipse.swt.widgets.*;
